Initial commit.
I haven't been using any form of source code management, so there's a whole lot of old/unnecessary crap in here. This is really just for archiving purposes, as I'm about to clean up and delete a lot of old code.

If you want to contribute, consider waiting until after I've cleaned up. It'll be a lot easier to figure out what's going on.