Parse automatch request parameters into numbers
Somehow dsc is generating automatch requests where the max and min
number of players is a String rather than an int.  It may have something
to do with unicode parsing, as he likes to use funny Unicode symbols in
his game titles.  But since I can't find the parsing problem so I'm
addressing the symptom instead.